Day 6

+++ src/challenge_result.rs
@@ -4,12 +4,12 @@ pub type ChallengeResult = Result<ChallengeSuccess, Box<dyn Error>>;
 
 #[derive(Debug, Copy, Clone)]
 pub struct ChallengeSuccess {
-    pub part1: i32,
-    pub part2: i32,
+    pub part1: u64,
+    pub part2: u64,
 }
 
 impl ChallengeSuccess {
-    pub fn new(part1: i32, part2: i32) -> Self {
+    pub fn new(part1: u64, part2: u64) -> Self {
         ChallengeSuccess { part1, part2 }
     }
 }
